feat: 🎸 split internal eth transfers into table
