+++ v2/src/modules/address/components/EthBalanceTabs/TableAllEthTransferRow.vue
@@ -1,86 +1,100 @@
 <template>
-    <app-table-row v-if="smAndDown" @click="showMoreDetails = !showMoreDetails" :color="showMoreDetails ? 'pillGrey' : 'transparent'">
-        <v-col cols="6">
-            <v-avatar :color="transferDirection.color" size="22">
-                <v-icon size="13" color="white"> {{ transferDirection.icon }} </v-icon>
-            </v-avatar>
-            <span class="ml-2">
-                {{ transferDirection.text }}
-            </span>
-        </v-col>
-        <v-col cols="6" class="text-right">{{ txValue.value }} ETH</v-col>
-        <v-col cols="6" class="text-info text-lowercase"> {{ timestamp }} {{ transferDirection.direction }} </v-col>
-        <v-col cols="6">
-            <div class="d-flex align-center justify-end">
-                <app-address-blockie :address="txAddress || ''" :size="6" class="mr-2 mr-sm-2" />
-                <app-transform-hash is-short is-blue :hash="eth.toCheckSum(txAddress)" :link="`/address/${txAddress}`" />
-            </div>
-        </v-col>
-        <template #expandable>
-            <v-col cols="12" v-if="showMoreDetails">
-                <div v-if="!isIncoming" class="mb-4 d-flex justify-space-between">
-                    <span class="text-info mr-2">tx fee paid:</span>
-                    <span>{{ txFee.value }} ETH</span>
-                </div>
-                <div class="d-flex justify-space-between">
-                    <span class="text-info mr-2">
-                        <app-chip :bg="transferType.color" :text="transferType.text.toUpperCase()" class="text-caption" style="min-width: 120px" />
-                    </span>
-                    <span>
-                        <app-transform-hash
-                            is-short
-                            is-blue
-                            :hash="eth.toCheckSum(transfer.transfer.transactionHash)"
-                            :link="`/tx/${transfer.transfer.transactionHash}`"
-                        />
-                    </span>
+    <div>
+        <app-table-row v-if="smAndDown" @click="showMoreDetails = !showMoreDetails" :color="showMoreDetails ? 'pillGrey' : 'transparent'">
+            <v-col cols="6">
+                <v-avatar :color="transferDirection.color" size="22">
+                    <v-icon size="13" color="white"> {{ transferDirection.icon }} </v-icon>
+                </v-avatar>
+                <span class="ml-2">
+                    {{ transferDirection.text }}
+                </span>
+            </v-col>
+            <v-col cols="6" class="text-right">{{ txValue.value }} ETH</v-col>
+            <v-col cols="6" class="text-info text-lowercase"> {{ timestamp }} {{ transferDirection.direction }} </v-col>
+            <v-col cols="6">
+                <div class="d-flex align-center justify-end">
+                    <app-address-blockie :address="txAddress || ''" :size="6" class="mr-2 mr-sm-2" />
+                    <app-transform-hash is-short is-blue :hash="eth.toCheckSum(txAddress)" :link="`/address/${txAddress}`" />
                 </div>
             </v-col>
-        </template>
-    </app-table-row>
-    <app-table-row v-else row-align="center">
-        <v-col sm="3" lg="2">
-            <div class="d-flex">
-                <span style="width: 30px" class="d-inline-block">
-                    <v-icon :color="statusIcon.color">{{ statusIcon.icon }}</v-icon>
-                </span>
-                <div class="ml-4">
-                    {{ txValue.value }} ETH
-                    <div class="text-lowercase">
-                        {{ timestamp }}
+            <template #expandable>
+                <v-col cols="12" v-if="showMoreDetails">
+                    <div v-if="!isIncoming" class="mb-4 d-flex justify-space-between">
+                        <span class="text-info mr-2">tx fee paid:</span>
+                        <span>{{ txFee.value }} ETH</span>
                     </div>
-                    <app-chip v-if="mdAndDown" :bg="transferType.color" :text="transferType.text.toUpperCase()" class="tiny-text" style="min-width: 120px" />
+                    <div class="d-flex justify-space-between">
+                        <span class="text-info mr-2">
+                            <app-chip :bg="transferType.color" :text="transferType.text.toUpperCase()" class="text-caption" style="min-width: 120px" />
+                        </span>
+                        <span>
+                            <app-transform-hash
+                                is-short
+                                is-blue
+                                :hash="eth.toCheckSum(transfer.transfer.transactionHash)"
+                                :link="`/tx/${transfer.transfer.transactionHash}`"
+                            />
+                        </span>
+                    </div>
+                </v-col>
+            </template>
+        </app-table-row>
+        <app-table-row v-else row-align="center">
+            <v-col sm="3" lg="2">
+                <div class="d-flex">
+                    <span style="width: 30px" class="d-inline-block">
+                        <v-icon :color="statusIcon.color">{{ statusIcon.icon }}</v-icon>
+                    </span>
+                    <div class="ml-4">
+                        {{ txValue.value }} ETH
+                        <div class="text-lowercase">
+                            {{ timestamp }}
+                        </div>
+                        <app-chip
+                            v-if="mdAndDown"
+                            :bg="transferType.color"
+                            :text="transferType.text.toUpperCase()"
+                            class="tiny-text"
+                            style="min-width: 120px"
+                        />
+                    </div>
                 </div>
-            </div>
-        </v-col>
-        <v-col v-if="!mdAndDown" sm="2">
-            <app-chip :bg="transferType.color" :text="transferType.text.toUpperCase()" size="x-small" class="tiny-text" style="min-width: 131px" />
-        </v-col>
-        <v-col sm="2">
-            <router-link v-if="isBlockReward" :to="`/block/number/${props.transfer.transfer.block}`" class="text-secondary">
-                {{ formatNumber(props.transfer.transfer.block) }}
-            </router-link>
-            <app-transform-hash
-                v-else
-                is-short
-                is-blue
-                :hash="eth.toCheckSum(props.transfer.transfer.transactionHash)"
-                :link="`/tx/${props.transfer.transfer.transactionHash}`"
-            />
-        </v-col>
-        <v-col sm="2">
-            <app-chip v-if="!isBlockReward" :bg="transferDirection.color" :text="transferDirection.direction" size="x-small" class="tiny-text text-uppercase" />
-        </v-col>
-        <v-col sm="3" lg="2">
-            <div v-if="!isBlockReward" class="d-flex align-center">
-                <app-address-blockie :address="txAddress || ''" :size="8" class="mr-2 mr-sm-2" />
-                <app-transform-hash is-short is-blue :hash="eth.toCheckSum(txAddress)" :link="`/address/${txAddress}`" />
-            </div>
-        </v-col>
-        <v-col sm="2">
-            <span v-if="!isIncoming"> {{ txFee.value }} ETH </span>
-        </v-col>
-    </app-table-row>
+            </v-col>
+            <v-col v-if="!mdAndDown" sm="2">
+                <app-chip :bg="transferType.color" :text="transferType.text.toUpperCase()" size="x-small" class="tiny-text" style="min-width: 131px" />
+            </v-col>
+            <v-col sm="2">
+                <router-link v-if="isBlockReward" :to="`/block/number/${props.transfer.transfer.block}`" class="text-secondary">
+                    {{ formatNumber(props.transfer.transfer.block) }}
+                </router-link>
+                <app-transform-hash
+                    v-else
+                    is-short
+                    is-blue
+                    :hash="eth.toCheckSum(props.transfer.transfer.transactionHash)"
+                    :link="`/tx/${props.transfer.transfer.transactionHash}`"
+                />
+            </v-col>
+            <v-col sm="2">
+                <app-chip
+                    v-if="!isBlockReward"
+                    :bg="transferDirection.color"
+                    :text="transferDirection.direction"
+                    size="x-small"
+                    class="tiny-text text-uppercase"
+                />
+            </v-col>
+            <v-col sm="3" lg="2">
+                <div v-if="!isBlockReward" class="d-flex align-center">
+                    <app-address-blockie :address="txAddress || ''" :size="8" class="mr-2 mr-sm-2" />
+                    <app-transform-hash is-short is-blue :hash="eth.toCheckSum(txAddress)" :link="`/address/${txAddress}`" />
+                </div>
+            </v-col>
+            <v-col sm="2">
+                <span v-if="!isIncoming"> {{ txFee.value }} ETH </span>
+            </v-col>
+        </app-table-row>
+    </div>
 </template>
 
 <script setup lang="ts">
