+++ v2/src/views/ViewAddressEthBalance.vue
@@ -8,7 +8,7 @@
                 <div>
                     <app-tabs v-model="state.tab" :routes="routes" :tabs="tabs" @update:modelValue="setLastViewedTab()" class="mb-4 mb-sm-0"></app-tabs>
                     <module-all-eth-transfers v-if="state.tab === routes[0]" :address-ref="props.addressRef" />
-                    <module-all-and-internal v-if="state.tab === routes[1]" :tab="state.tab" :address-ref="props.addressRef" />
+                    <module-internal-eth-transfers v-if="state.tab === routes[1]" :tab="state.tab" :address-ref="props.addressRef" />
                     <module-eth-txs-history v-if="state.tab === routes[2]" :address-ref="props.addressRef" />
                     <module-pending-transfers v-if="state.tab === routes[3]" :tab="state.tab" :address-ref="props.addressRef" />
                     <template v-if="props.isAddressMiner && state.tab === routes[4]">
@@ -23,7 +23,7 @@
 <script setup lang="ts">
 import AppTabs from '@/core/components/AppTabs'
 import ModulePendingTransfers from '@module/address/ModulePendingEthTransfers.vue'
-import ModuleAllAndInternal from '@module/address/ModuleAllAndInternalEthTransfers.vue'
+import ModuleInternalEthTransfers from '@module/address/ModuleInternalEthTransfers.vue'
 import ModuleAddressBalance from '@module/address/ModuleAddressBalance.vue'
 import ModuleEthTxsHistory from '@module/address/ModuleEthTxsHistory.vue'
 import ModuleAddressMinerBlock from '@module/address/ModuleAddressMinerBlock.vue'
